slice to module
