+++ polars/polars-lazy/src/physical_plan/expressions/mod.rs
@@ -3,11 +3,11 @@ pub(crate) mod alias;
 pub(crate) mod binary;
 pub(crate) mod column;
 pub mod default;
-mod final_agg;
 pub(crate) mod is_not_null;
 pub(crate) mod is_null;
 pub(crate) mod literal;
 pub(crate) mod not;
+pub(crate) mod slice;
 pub(crate) mod sort;
 pub(crate) mod sortby;
 pub(crate) mod take;
